requirements.txt
apache-airflow==1.10.2
alembic==0.8.10
appdirs==1.4.3
asn1crypto==1.2.0
azure-common==1.1.23
azure-nspkg==3.0.2
azure-storage==0.36.0
backoff==1.10.0
bcrypt==3.1.7
boto3==1.10.38
botocore==1.13.38
cached-property==1.5.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.0
configparser==3.5.3
croniter==0.3.30
cryptography==2.8
defusedxml==0.6.0
dill==0.2.9
docker==4.1.0
docker-compose==1.25.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
Flask==0.12.4
Flask-Admin==1.5.2
Flask-Cache==0.13.1
Flask-Login==0.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
future==0.16.0
gitdb2==2.0.6
GitPython==3.0.5
gnupg==2.3.1
gunicorn==19.10.0
idna==2.8
ijson==2.5.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
jsonschema==3.2.0
lockfile==0.12.2
lxml==4.4.2
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.1
meld3==2.0.0
mock==3.0.5
numpy==1.17.4
objectpath==0.6.1
ordereddict==1.1
pandas==0.25.3
paramiko==2.7.1
pbr==5.4.4
psutil==5.6.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodomex==3.9.4
Pygments==2.5.2
PyJWT==1.7.1
PyNaCl==1.3.0
pyodbc==4.0.27
pyOpenSSL==19.1.0
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python-gnupg==0.4.5
python-nvd3==0.15.0
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.1.2
psycopg2==2.8.4
requests==2.22.0
redis==3.3.11
requests-toolbelt==0.9.1
Salesforce-FuelSDK==1.3.0
s3transfer==0.2.1
setproctitle==1.1.10
simplejson==3.17.0
singer-python==5.9.0
six==1.13.0
smmap2==2.0.5
snowflake-connector-python==2.1.1
snowflake-sqlalchemy==1.1.17
SQLAlchemy==1.2.19
stripe==2.41.0
supervisor==4.1.0
tabulate==0.8.2
target-stitch==3.0.1
texttable==1.6.2
thrift==0.13.0
Unidecode==1.1.1
urllib3==1.25.7
websocket-client==0.56.0
Werkzeug==0.14.1
WTForms==2.2.1
zeep==3.4.0
zope.deprecation==4.4.0